Move node tests to a new directory
