+++ test/SystemTextJson.JsonDiffPatch.UnitTests/NodeTests/JsonValueComparerTests.cs
@@ -2,7 +2,7 @@
 using System.Text.Json.Nodes;
 using Xunit;
 
-namespace SystemTextJson.JsonDiffPatch.UnitTests
+namespace SystemTextJson.JsonDiffPatch.UnitTests.NodeTests
 {
     public class JsonValueComparerTests
     {
