+++ test/SystemTextJson.JsonDiffPatch.UnitTests/NodeTests/PatchTests.cs
@@ -3,7 +3,7 @@ using System.Text.Json.JsonDiffPatch;
 using System.Text.Json.Nodes;
 using Xunit;
 
-namespace SystemTextJson.JsonDiffPatch.UnitTests
+namespace SystemTextJson.JsonDiffPatch.UnitTests.NodeTests
 {
     public class PatchTests
     {
